json replacer and reviver
a bit hacky but handles Object descendants such as RegExp and Date.  For
the replacer it will defer to a toJSON method if it exists, the reviver
alas must hack through ever string to see if it is something generated
by toJSON.  I am open to ideas on this one.